+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-## [0.4.0] - 2020-07-14
 ### Added
 - Summary collector
+
+### Deprecated
+- Average collector
 
 ## [0.4.0] - 2020-07-14
 ### Added